ml-service/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
joblib==1.4.2